refactor: reuse initial-pr content for subgroup prs
